+++ src/main/java/roomescape/client/dto/TossErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.client.dto;
+package roomescape.client.dto;
 
 public record TossErrorResponse(String message) {
 }
