+++ src/main/java/roomescape/payment/repository/PaymentRepository.java
@@ -1,8 +1,8 @@
-package roomescape.reservation.repository;
+package roomescape.payment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.reservation.domain.Payment;
-import roomescape.reservation.domain.PaymentKey;
+import roomescape.payment.domain.Payment;
+import roomescape.payment.domain.PaymentKey;
 
 public interface PaymentRepository extends JpaRepository<Payment, PaymentKey> {
 }
